fix: useEventOperations.spec.ts server.use return 에러 수정
